requirements.txt
Django>=3.2,<4.2
djongo>=1.3.6
graphene-django>=3.2
pillow
prisma
stripe
django-storages[s3]
boto3
gunicorn
